Move StdTimeout to ectool under std feature
